add func to send email
